requirements.txt
ag2==0.9.7
autogen-ext==0.7.2
anthropic==0.64.0
streamlit>=1.36
python-dotenv>=1.0

# Enhanced Business Intelligence Platform Dependencies
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
markdown2>=2.4.0

# Database and Storage
# sqlite3 is built into Python - no installation needed
psycopg2-binary>=2.9.0  # PostgreSQL adapter for production
redis>=4.5.0  # Redis client for caching and sessions

# Vector databases for RAG functionality
chromadb>=0.4.0  # Primary vector database
faiss-cpu>=1.7.0  # Facebook AI Similarity Search
# pinecone-client>=2.2.0  # Optional: Pinecone cloud vector DB
# openai>=1.0.0  # Optional: If using OpenAI models alongside Claude

# Web scraping and document processing
beautifulsoup4>=4.12.0
lxml>=4.9.0
html5lib>=1.1

# Data processing and scientific computing
scikit-learn>=1.3.0
scipy>=1.11.0

# System monitoring and health checks
psutil>=5.9.0  # System and process monitoring

# Development and Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0  # Coverage reporting
pytest-xdist>=3.3.0  # Parallel test execution
pytest-mock>=3.11.0  # Enhanced mocking
coverage>=7.3.0  # Coverage analysis